requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-datetime-widget==0.9.3
django-stdimage==2.4.2
idna==2.6
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
progressbar2==3.34.2
python-utils==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
psycopg2
